+++ pdfalyzer/helpers/pdf_object_helper.py
@@ -2,59 +2,103 @@
 Some methods to help with the direct manipulation/processing of PyPDF2's PdfObjects
 """
 from collections import namedtuple
-from typing import List
+from typing import List, Optional
 
 from PyPDF2.generic import IndirectObject, PdfObject
 from rich.markup import escape
+from yaralyzer.helpers.string_helper import comma_join
 from yaralyzer.util.logging import log
 
 from pdfalyzer.output.layout import get_label_style
 from pdfalyzer.util.adobe_strings import *
 from pdfalyzer.util.exceptions import PdfWalkError
 
-# In the case of easy key/value pairs the reference_key and the reference_address are the same but
-# for more complicated references the reference_address will be the reference_key plus sub references.
-#
-#   e.g. a reference_key for a /Font labeled /F1 could be '/Resources' but the reference_address
-# might be '/Resources[/Font][/F1] if the /Font is a directly embedded reference instead of a remote one.
-PdfObjectRef = namedtuple('PdfObjectRef', ['reference_key', 'reference_address', 'pdf_obj'])
-
 # For printing SymlinkNodes
 SymlinkRepresentation = namedtuple('SymlinkRepresentation', ['text', 'style'])
 
 
-def get_references(obj: PdfObject, ref_key=None, ref_address=None) -> List[PdfObjectRef]:
-    """Return list of PdfObjectRefs"""
-    if isinstance(obj, IndirectObject):
-        if ref_key is None or ref_address is None:
-            raise PdfWalkError(f"{obj} is a reference but key or address not provided")
+class PdfObjectRelationship:
+    """
+    Simple container class for information about a link between two PDF objects.
+    In the case of easy key/value pairs the reference_key and the reference_address are the same but
+    for more complicated references the reference_address will be the reference_key plus sub references.
 
-        return [PdfObjectRef(ref_key, ref_address, obj)]
-    elif isinstance(obj, list) and isinstance(obj, dict):
-        raise PdfWalkError(f"PdfObject {obj} is both a dict and a list")
+    e.g. a reference_key for a /Font labeled /F1 could be '/Resources' but the reference_address
+    might be '/Resources[/Font][/F1] if the /Font is a directly embedded reference instead of a remote one.
+    """
+    def __init__(
+            self,
+            from_obj: PdfObject,
+            to_obj: IndirectObject,
+            reference_key: str,
+            reference_address: str
+        ) -> None:
+        self.from_obj = from_obj
+        self.to_obj = to_obj
+        self.reference_key = reference_key
+        self.reference_address = reference_address
+        self.from_node: Optional['PdfTreeNode'] = None  # To be filled in later.  TODO: Hacky
 
-    return_list = []
+    @classmethod
+    def get_references(
+            cls,
+            obj: PdfObject,
+            ref_key: Optional[str] = None,
+            ref_address: Optional[str] = None
+        ) -> List['PdfObjectRelationship']:
+        """
+        Recurse through elements in 'obj' and return list of PdfObjectRelationships containing all IndirectObjects
+        referenced from addresses in 'obj'.
+        """
+        if isinstance(obj, IndirectObject):
+            if ref_key is None or ref_address is None:
+                raise PdfWalkError(f"{obj} is a reference but key or address not provided")
+            else:
+                return [cls(obj, obj, ref_key, ref_address)]
 
-    if isinstance(obj, list):
-        for i, element in enumerate(obj):
-            if not isinstance(element, (IndirectObject, list, dict)):
-                continue
+        return_list: List[PdfObjectRelationship] = []
 
-            _ref_address = f"[{i}]" if ref_address is None else f"{ref_address}[{i}]"
-            return_list += get_references(element, ref_key or i, _ref_address)
-    elif isinstance(obj, dict):
-        for k, v in obj.items():
-            _ref_address = k if ref_address is None else f"{ref_address}[{k}]"
-            return_list += get_references(v, ref_key or k, _ref_address)
-    else:
-        log.debug(f"Adding no references for PdfObject reference '{ref_key}' -> '{obj}'")
+        if isinstance(obj, list):
+            for i, element in enumerate(obj):
+                if not isinstance(element, (IndirectObject, list, dict)):
+                    continue
 
-    return return_list
+                _ref_address = f"[{i}]" if ref_address is None else f"{ref_address}[{i}]"
+                return_list += cls.get_references(element, ref_key or i, _ref_address)
+        elif isinstance(obj, dict):
+            for k, v in obj.items():
+                _ref_address = k if ref_address is None else f"{ref_address}[{k}]"
+                return_list += cls.get_references(v, ref_key or k, _ref_address)
+        else:
+            log.debug(f"Adding no references for PdfObject reference '{ref_key}' -> '{obj}'")
+
+        for ref in return_list:
+            ref.from_obj = obj
+
+        return return_list
+
+    def __eq__(self, other: 'PdfObjectRelationship') -> bool:
+        """Note that equality does not check self.from_obj."""
+        if (self.to_obj.idnum != other.to_obj.idnum) or (self.from_node != other.from_node):
+            return False
+
+        for k in ['reference_key', 'reference_address']:
+            if getattr(self, k) != getattr(other, k):
+                return False
+
+        return True
+
+    def __str__(self) -> str:
+        return comma_join([f"{k}: {v}" for k, v in vars(self).items()])
+
+    def description(self) -> str:
+        """Sort of like __str__ but w/out the extra lines"""
+        return f"{self.from_node}: {self.reference_address} to {self.to_obj}"
 
 
 def get_symlink_representation(from_node, to_node) -> SymlinkRepresentation:
     """Returns a tuple (symlink_text, style) that can be used for pretty printing, tree creation, etc"""
-    reference_key = str(to_node.get_reference_key_for_relationship(from_node))
+    reference_key = str(to_node.get_address_for_relationship(from_node))
     pdf_instruction = reference_key.split('[')[0]  # In case we ended up with a [0] or similar
 
     if pdf_instruction in DANGEROUS_PDF_KEYS:
@@ -75,3 +119,7 @@ def pdf_object_id(pdf_object):
 def does_list_have_any_references(_list) -> bool:
     """Return true if any element of _list is an IndirectObject"""
     return any(isinstance(item, IndirectObject) for item in _list)
+
+
+def _sort_pdf_object_refs(refs: List[PdfObjectRelationship]) -> List[PdfObjectRelationship]:
+    return sorted(refs, key=lambda ref: ref.to_obj.idnum)